test(sync_manager): split E2E tests by CRDT family
