+++ crates/nx-core/src/sync_manager/tests/e2e/mod.rs
@@ -1,0 +1,6 @@
+mod gcounter;
+mod lww_map;
+mod lww_register;
+mod orset;
+mod pncounter;
+mod rga;
